Refactor emandate html buildup into library
