+++ cyclos_configuration/scripts/emandates/Library.groovy
@@ -9,9 +9,13 @@ import org.cyclos.entities.system.CustomOperationField
 import org.cyclos.entities.system.CustomWizardExecution
 import org.cyclos.entities.system.CustomWizardField
 import org.cyclos.entities.system.ExternalRedirectExecution
+import org.cyclos.entities.users.QRecordCustomFieldPossibleValue
+import org.cyclos.entities.users.RecordCustomFieldPossibleValue
 import org.cyclos.entities.users.SystemRecord
 import org.cyclos.entities.users.SystemRecordType
 import org.cyclos.entities.users.User
+import org.cyclos.impl.access.ConfigurationHandler
+import org.cyclos.impl.contentmanagement.DataTranslationHandler
 import org.cyclos.impl.InvocationContext
 import org.cyclos.impl.system.BaseCustomFieldPossibleValueCategoryServiceLocal
 import org.cyclos.impl.system.BaseCustomFieldPossibleValueServiceLocal
@@ -25,6 +29,7 @@ import org.cyclos.impl.system.ScriptHelper
 import org.cyclos.impl.users.RecordServiceLocal
 import org.cyclos.impl.users.UserServiceLocal
 import org.cyclos.impl.utils.LinkGeneratorHandler
+import org.cyclos.impl.utils.formatting.FormatterImpl
 import org.cyclos.impl.utils.persistence.EntityManagerHandler
 import org.cyclos.model.ValidationException
 import org.cyclos.model.system.fields.CustomFieldPossibleValueCategoryDTO
@@ -60,6 +65,9 @@ class EMandates {
 	
 	ObjectMapper objectMapper
 	ScriptHelper scriptHelper
+	ConfigurationHandler configurationHandler
+	DataTranslationHandler dataTranslationHandler
+	FormatterImpl formatter
 	ServletContext servletContext
 	EntityManagerHandler entityManagerHandler
 	CustomWizardFieldPossibleValueServiceLocal wizardPossibleValueService
@@ -75,6 +83,9 @@ class EMandates {
 	EMandates(Binding binding) {
 		def vars = binding.variables
 		scriptHelper = vars.scriptHelper as ScriptHelper
+		configurationHandler = vars.configurationHandler as ConfigurationHandler
+		dataTranslationHandler = vars.dataTranslationHandler as DataTranslationHandler
+		formatter = vars.formatter as FormatterImpl
 		objectMapper = vars.objectMapper as ObjectMapper
 		entityManagerHandler = vars.entityManagerHandler as EntityManagerHandler
 		wizardPossibleValueService = vars.customWizardFieldPossibleValueService as CustomWizardFieldPossibleValueServiceLocal
@@ -484,5 +495,61 @@ class EMandates {
 		return "A total of ${records.size()} eMandates were pending, of which " + 
 			"${nowSuccess} are now successful, ${nowExpired} are now expired " +
 			" and ${stillPending} are still pending"
+	}
+
+	/**
+	 * Builds up an HTML string containing relevant details from the given emandate record.
+	 */
+	String emandateHtml(SystemRecord record, User user) {
+		if ( ! record ) {
+			return ''
+		}
+		def fields = scriptHelper.wrap(record)
+		def config = configurationHandler.getAccessor(user)
+
+		def html = ''
+		def details = scriptParameters["details"]
+		html += "<div>${details}</div>"
+
+		def bankName = fields.bankName
+		def bankNameField = record.type.fields.find { it.internalName == 'bankName' }
+		def bankNameLabel = dataTranslationHandler.getName(config.language, bankNameField)
+		html += "<div><strong>${bankNameLabel}:</strong> ${bankName}</div>"
+
+		def iban = fields.iban
+		if (iban) {		
+			def ibanField = record.type.fields.find { it.internalName == 'iban' }
+			def ibanLabel = dataTranslationHandler.getName(config.language, ibanField)
+			html += "<div><strong>${ibanLabel}:</strong> ${iban}</div>"
+		}
+
+		def accountName = fields.accountName
+		if (accountName) {
+			def accountNameField = record.type.fields.find { it.internalName == 'accountName' }
+			def accountNameLabel = dataTranslationHandler.getName(config.language, accountNameField)
+			html += "<div><strong>${accountNameLabel}:</strong> ${accountName}</div>"
+		}
+
+		def signerName = fields.signerName
+		if (signerName) {
+			def signerNameField = record.type.fields.find { it.internalName == 'signerName' }
+			def signerNameLabel = dataTranslationHandler.getName(config.language, signerNameField)
+			html += "<div><strong>${signerNameLabel}:</strong> ${signerName}</div>"
+		}
+
+		def statusField = record.type.fields.find { it.internalName == 'status' }
+		def statusLabel = dataTranslationHandler.getName(config.language, statusField)
+		RecordCustomFieldPossibleValue statusValue = fields.status as RecordCustomFieldPossibleValue
+		def pv = QRecordCustomFieldPossibleValue.recordCustomFieldPossibleValue
+		def statusValueLabel = dataTranslationHandler.getValue(config.language, statusValue, pv.value)
+		html += "<div><strong>${statusLabel}:</strong> ${statusValueLabel}</div>"
+		
+		def statusDate = fields.statusDate
+		if (statusDate) {
+			def statusDateField = record.type.fields.find { it.internalName == 'statusDate' }
+			def statusDateLabel = dataTranslationHandler.getName(config.language, statusDateField)
+			html += "<div><strong>${statusDateLabel}:</strong> ${formatter.format(statusDate)}</div>"
+		}
+		return html
 	}
 }
